requirements.txt
streamlit
opencv-python
pytesseract
Pillow
numpy
easyocr
